Create a login error component for the home page
