+++ src/pages/Home.tsx
@@ -10,43 +10,23 @@ import {
 import { authenticatedRoutes } from 'app/routes';
 import HeroImageAndDescription from 'components/hero';
 import Dashboard from 'components/home/dashboard';
-import LoginNotifications from 'components/login/Notifications';
+import LoginError from 'components/home/LoginError';
 import { useShowDashboardWelcome } from 'context/DashboardWelcome';
 import { defaultOutline, semiTransparentBackground } from 'context/Theme';
-import useGlobalSearchParams, {
-    GlobalSearchParams,
-} from 'hooks/searchParams/useGlobalSearchParams';
 import usePageTitle from 'hooks/usePageTitle';
 import { Plus, Xmark } from 'iconoir-react';
-import { useMemo } from 'react';
 import { FormattedMessage } from 'react-intl';
 
 const Home = () => {
     const theme = useTheme();
 
     usePageTitle({ header: authenticatedRoutes.home.title });
-    const homePageError = useGlobalSearchParams(
-        GlobalSearchParams.HOME_PAGE_ERROR
-    );
 
     const { welcomeShown, setWelcomeShown } = useShowDashboardWelcome();
 
-    // home page error related
-    const notificationMessage = useMemo(() => {
-        if (homePageError) {
-            return 'directives.grant.skipped.message';
-        }
-        return null;
-    }, [homePageError]);
-
     return (
         <>
-            {notificationMessage ? (
-                <LoginNotifications
-                    notificationMessage={notificationMessage}
-                    notificationTitle="directives.grant.skipped.title"
-                />
-            ) : null}
+            <LoginError />
 
             <Box
                 sx={{
